pyproject.toml
[tool.poetry]
name = "ci-project"
version = "0.1.0"
description = ""
authors = ["SW <aliwo@naver.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.85.1"
uvicorn = "^0.19.0"
tortoise-orm = "^0.19.2"
aiomysql = "^0.1.1"
aerich = "^0.7.1"

[tool.poetry.group.dev.dependencies]
black = "^22.10.0"
mypy = "^0.982"
isort = "^5.10.1"
coverage = "^6.5.0"
pytest = "^7.1.3"
pytest-asyncio = "^0.20.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120

[tool.isort]
profile = "black"

[tool.mypy]
# Ref. https://mypy.readthedocs.io/en/stable/config_file.html
python_version = "3.10"
strict = true
plugins = "pydantic.mypy"
disallow_subclassing_any = false
ignore_missing_imports = true
implicit_reexport = true
show_error_codes = true
